+++ app/src/main/java/me/drakeet/meizhi/ui/adapter/MeizhiListAdapter.java
@@ -33,7 +33,7 @@ import com.squareup.picasso.Picasso;
 import java.io.IOException;
 import java.util.List;
 import me.drakeet.meizhi.R;
-import me.drakeet.meizhi.event.OnMeizhiTouchListener;
+import me.drakeet.meizhi.func.OnMeizhiTouchListener;
 import me.drakeet.meizhi.model.Meizhi;
 import me.drakeet.meizhi.widget.RatioImageView;
 import rx.Observable;
